+++ test/unit/stats/rtpStats.test.ts
@@ -1,4 +1,5 @@
-import { handleChromeStats, handleFirefoxSafariStats } from '../../../lib/stats/rtpStats';
+import { resolve } from 'path';
+import { handleWebRTCStats } from '../../../lib/stats/rtpStats';
 import { getChromeStatsResponse, getFirefoxSafariStatsResponse } from '../../mock/RTCStatsResponse';
 import rtpStatsResponse from '../../payloads/rtpStatsEvent.json';
 
@@ -50,39 +51,50 @@ describe('RTPStats', () => {
         };
     });
 
-    it('should validate chrome stats', () => {
+    it('should validate chrome stats', async () => {
         updateChromeContext(context);
-        handleChromeStats.call(context, stream);
+        handleWebRTCStats.call(context, stream);
         // call second time for collecting basepackets
-        handleChromeStats.call(context, stream);
-        expect(context.collected).toStrictEqual(rtpStatsResponse);
+        handleWebRTCStats.call(context, stream);
+        await new Promise<void>(res => setTimeout(() => {
+          let expected = JSON.parse(JSON.stringify(rtpStatsResponse));
+          expect(context.collected).toStrictEqual(expected);
+          res();
+      }, 100))    
     });
 
-    it('should validate chrome stats when enableQualityTracking flag is set to remoteonly', () => {
+    it('should validate chrome stats when enableQualityTracking flag is set to remoteonly', async () => {
         updateChromeContext(context);
         context.options.enableQualityTracking = "REMOTEONLY";
-        handleChromeStats.call(context, stream);
+        handleWebRTCStats.call(context, stream);
         // call second time for collecting basepackets
-        handleChromeStats.call(context, stream);
-        expect(context.collected).toStrictEqual(rtpStatsResponse);
+        handleWebRTCStats.call(context, stream);
+        await new Promise<void>(res => setTimeout(() => {
+          let expected = JSON.parse(JSON.stringify(rtpStatsResponse));
+          expect(context.collected).toStrictEqual(expected);
+          res();
+      }, 100))    
     });
 
-    it('should validate chrome stats when enableQualityTracking flag is default and enableTraking is false', () => {
+    it('should validate chrome stats when enableQualityTracking flag is default and enableTraking is false', async () => {
         updateChromeContext(context);
         context.options.enableTracking = false;
-        handleChromeStats.call(context, stream);
+        handleWebRTCStats.call(context, stream);
         // call second time for collecting basepackets
-        handleChromeStats.call(context, stream);
+        handleWebRTCStats.call(context, stream);
         console.log(context.collected);
-        expect(context.collected).toStrictEqual(rtpStatsResponse);
-    });
+        await new Promise<void>(res => setTimeout(() => {
+          let expected = JSON.parse(JSON.stringify(rtpStatsResponse));
+          expect(context.collected).toStrictEqual(expected);
+          res();
+      }, 100))    });
 
 
     it('should validate firefox stats', async () => {
         updateFirefoxContext(context);
-        handleFirefoxSafariStats.call(context, stream);
+        handleWebRTCStats.call(context, stream);
         // call second time for collecting basepackets
-        handleFirefoxSafariStats.call(context, stream);
+        handleWebRTCStats.call(context, stream);
         await new Promise<void>(res => setTimeout(() => {
             let expected = JSON.parse(JSON.stringify(rtpStatsResponse));
             expected.local.rtt = 0.026;
@@ -103,9 +115,9 @@ describe('RTPStats', () => {
 
   it('should validate firefox stats', async () => {
     updateFirefoxContext(context);
-    handleFirefoxSafariStats.call(context, stream);
+    handleWebRTCStats.call(context, stream);
     // call second time for collecting basepackets
-    handleFirefoxSafariStats.call(context, stream);
+    handleWebRTCStats.call(context, stream);
     await new Promise<void>((res) => setTimeout(() => {
       const expected = JSON.parse(JSON.stringify(rtpStatsResponse));
       expected.local.rtt = 0.026;
@@ -136,7 +148,7 @@ describe('RTPStats', () => {
         sdp: 'a=rtpmap:0 opus/8000',
       },
     };
-    handleFirefoxSafariStats.call(context, stream);
+    handleWebRTCStats.call(context, stream);
     await new Promise<void>((res) => setTimeout(() => {
       expect(consoleSpy.mock.calls[0][1]).toMatch(/Error in getStats LocalStreams API/);
       res();
@@ -145,12 +157,12 @@ describe('RTPStats', () => {
 
   it('should validate safari stats', async () => {
     updateSafariContext(context);
-    handleFirefoxSafariStats.call(context, stream);
+    handleWebRTCStats.call(context, stream);
     // call second time for collecting basepackets
     stream = {
       codec: '', local: {}, remote: {}, networkType: '',
     };
-    handleFirefoxSafariStats.call(context, stream);
+    handleWebRTCStats.call(context, stream);
     await new Promise<void>((res) => setTimeout(() => {
       const expected = JSON.parse(JSON.stringify(rtpStatsResponse));
       expected.codec = 'pcmu';
@@ -172,10 +184,12 @@ describe('RTPStats', () => {
 
 const updateChromeContext = (context: any) => {
   context.pc = {
-    getStats: (cb, selector, errback) => {
-      cb(getChromeStatsResponse());
+    getSenders: chromeGetStatsMock,
+    getReceivers: chromeGetStatsMock,
+    remoteDescription: {
+      sdp: 'a=rtpmap:0 opus/8000',
     },
-  };
+  }
   context.clientScope.browserDetails.browser = 'chrome';
 };
 
@@ -204,5 +218,11 @@ const updateSafariContext = (context: any) => {
 const ffSafariGetStatsMock = () => [
   {
     getStats: () => Promise.resolve(getFirefoxSafariStatsResponse()),
+  },
+];
+
+const chromeGetStatsMock = () => [
+  {
+    getStats: () => Promise.resolve(getChromeStatsResponse()),
   },
 ];